ganti atribut time player
Time dibikin float dlu,.. ntar saat diprint di jadiin int

+++ lib/player.h
@@ -14,7 +14,7 @@
 typedef struct
 {
     LOCATION currentLocation;
-    int time;
+    float time;
     int money;
     int speed;
     int bagCapacity;
